Deeper scan part 4: Convert pandas/numpy values to float in ML/RL/feature engineering modules
- investment_recommender.py: Cast current_price, volume, change to float
- multi_source_data.py: Cast FRED and Yahoo macro indicators to float
- momentum_predictor.py: Cast ROC, momentum, velocity to float
- backtester.py: Cast final_price, equity_curve metrics to float
- ml_predictor.py: Cast all technical indicator features to float
- ensemble_predictor.py: Cast all features including CCI, Williams %R, Stochastic to float
- deep_learning_predictor.py: Cast all LSTM features to float
- rl_trader.py: Cast all RL state features to float
- online_learner.py: Cast all online learning features to float
- auto_feature_engineer.py: Cast all base features to float
- ml_backtester.py: Cast buy-hold return to float

All numeric values now explicitly cast to float with empty DataFrame checks to prevent JSON serialization errors and frontend .toFixed issues.

+++ investment_recommender.py
@@ -497,11 +497,11 @@ class InvestmentRecommender:
                 if hist.empty:
                     return None
 
-                current_price = hist['Close'].iloc[-1]
+                current_price = float(hist['Close'].iloc[-1])
                 return {
                     'current_price': current_price,
-                    'volume': hist['Volume'].iloc[-1],
-                    'change': (current_price - hist['Close'].iloc[0]) / hist['Close'].iloc[0],
+                    'volume': float(hist['Volume'].iloc[-1]),
+                    'change': float((current_price - hist['Close'].iloc[0]) / hist['Close'].iloc[0]),
                     'info': info
                 }
